Ensure size_t is used for each storage::IStorage implementation

+++ components/ocs_control/test/test_fsm_block_storage.h
@@ -19,9 +19,9 @@ public:
                         status::StatusCode write_status = status::StatusCode::OK,
                         status::StatusCode erase_status = status::StatusCode::OK);
 
-    status::StatusCode probe(const char* key, unsigned& size) override;
-    status::StatusCode read(const char* key, void* data, unsigned size) override;
-    status::StatusCode write(const char* key, const void* data, unsigned size) override;
+    status::StatusCode probe(const char* key, size_t& size) override;
+    status::StatusCode read(const char* key, void* data, size_t size) override;
+    status::StatusCode write(const char* key, const void* data, size_t size) override;
     status::StatusCode erase(const char* key) override;
 
     FsmBlock::State prev_state { 0 };
